requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==0.16
Flask-Login==0.1.3
Flask-Evolution==0.6
Flask-Script==0.3.3
Jinja2==2.6
Werkzeug==0.8.3
wsgiref==0.1.2
pysqlite==2.6.3

flask-openid
flask-mail==0.7.6
sqlalchemy==0.7.9
sqlalchemy-migrate
flask-whooshalchemy==0.54a
flask-wtf==0.8.4
pytz==2013b
flask-babel==0.8
flup